requirements.txt
numpy
pandas
torch
pyyaml
colorama